requirements.txt
python-dotenv
streamlit
langchain-community
langchain-text-splitters
langchain-google-genai
unstructured
unstructured[pdf]
faiss-cpu
transformers
sentence-transformers
nltk